backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.7.0
python-dotenv==1.0.1
huggingface-hub==0.27.1
requests==2.32.3
